# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Earning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{telegramId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{telegramId}/unban:
    post:
      operationId: unbanUser
      tags: [admin]
      summary: Unban a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List all withdrawal requests (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWithdrawalList"
  /admin/withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject withdrawal
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/broadcast:
    post:
      operationId: adminBroadcast
      tags: [admin]
      summary: Send broadcast notification to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get all admin settings
      responses:
        "200":
          description: Admin settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /admin/force-join:
    get:
      operationId: listForceJoinChannels
      tags: [admin]
      summary: List force join channels
      responses:
        "200":
          description: Force join channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForceJoinChannel"
    post:
      operationId: addForceJoinChannel
      tags: [admin]
      summary: Add a force join channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForceJoinChannelInput"
      responses:
        "201":
          description: Channel added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForceJoinChannel"
  /admin/force-join/{channelId}:
    delete:
      operationId: deleteForceJoinChannel
      tags: [admin]
      summary: Delete a force join channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/activity-logs:
    get:
      operationId: getActivityLogs
      tags: [admin]
      summary: Get admin activity/audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogList"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        coins:
          type: integer
        totalEarnings:
          type: integer
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        rank:
          type: string
        isBanned:
          type: boolean
        dailyBonusLastClaimed:
          type: ["string", "null"]
        streakDays:
          type: integer
        tasksCompleted:
          type: integer
        createdAt:
          type: string
        lastActive:
          type: string
      required: [id, telegramId, firstName, coins, totalEarnings, referralCode, rank, isBanned, streakDays, tasksCompleted, createdAt, lastActive]
...
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        telegramId:
          type: string
        amount:
          type: integer
        method:
          type: string
        accountDetails:
          type: string
        status:
          type: string
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, telegramId, amount, method, accountDetails, status, createdAt, updatedAt]
...
    ForceJoinChannel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        url:
          type: string
        type:
          type: string
        isActive:
          type: boolean
      required: [id, name, username, url, type, isActive]
    ForceJoinChannelInput:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        url:
          type: string
        type:
          type: string
      required: [name, username, url, type]
...
    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AdminLoginResult:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
      required: [success, token, admin]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        lastLogin:
          type: ["string", "null"]
      required: [id, username, role]
...
    ActivityEntry:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        telegramId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, description, createdAt]
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [users, total, page, limit]
    BanInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...
    AdminWithdrawalList:
      type: object
      properties:
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
      required: [withdrawals, total, page]
    RejectInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    BroadcastInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
      required: [title, message, type]
    AdminSettings:
      type: object
      properties:
        dailyBonusAmount:
          type: integer
        referralBonusAmount:
          type: integer
        minWithdrawalAmount:
          type: integer
        maxWithdrawalAmount:
          type: integer
        botUsername:
          type: string
        maintenanceMode:
          type: boolean
        withdrawMethods:
          type: array
          items:
            type: string
      required: [dailyBonusAmount, referralBonusAmount, minWithdrawalAmount, maxWithdrawalAmount, botUsername, maintenanceMode, withdrawMethods]
    AdminSettingsUpdate:
      type: object
      properties:
        dailyBonusAmount:
          type: integer
        referralBonusAmount:
          type: integer
        minWithdrawalAmount:
          type: integer
        maxWithdrawalAmount:
          type: integer
        botUsername:
          type: string
        maintenanceMode:
          type: boolean
        withdrawMethods:
          type: array
          items:
            type: string
    ActivityLogList:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
        total:
          type: integer
        page:
          type: integer
      required: [logs, total, page]
...